safari-600.7-branch/LayoutTests/http/tests/w3c/webperf/approved/navigation-timing/html/test_performance_attributes_exist_in_object.html

<!DOCTYPE html>
<html>
    <head>
        <meta charset="utf-8" />
        <title>window.performance attribute exists in an object</title>
        <link rel="author" title="Google" href="http://www.google.com/" />
	<link rel="help" href="http://www.w3.org/TR/navigation-timing/#sec-window.performance-attribute"/>
        <script src="/w3c/resources/testharness.js"></script>
        <script src="/w3c/resources/testharnessreport.js"></script>
        <script src="/w3c/webperf/resources/webperftestharness.js"></script>
<script id="metadata_cache">/*
{
  "window.performance is defined": {
    "help": "http://www.w3.org/TR/navigation-timing/#sec-window.performance-attribute",
    "assert": "The window.performance attribute provides a hosting area for performance related attributes. ",
    "author": "W3C http://www.w3.org/"
  },
  "window.performance.timing is defined": {
    "help": "http://www.w3.org/TR/navigation-timing/#sec-window.performance-attribute",
    "assert": "The window.performance attribute provides a hosting area for performance related attributes. ",
    "author": "W3C http://www.w3.org/"
  },
  "window.performance.navigation is defined": {
    "help": "http://www.w3.org/TR/navigation-timing/#sec-window.performance-attribute",
    "assert": "The window.performance attribute provides a hosting area for performance related attributes. ",
    "author": "W3C http://www.w3.org/"
  }
}
*/</script>
        <link rel="stylesheet" href="/w3c/resources/testharness.css" />
        <script>
          function onload_test() {
            var objWindow =
                document.getElementById("objectContext").contentDocument.defaultView;
            performanceNamespace = objWindow.performance;
            test_namespace('timing');
            test_namespace('navigation', true);
          }
        </script>
    </head>
    <body>
        <h1>Description</h1>
        <p>This test validates that the window.performance object exists in an
        object element (but does not validate that their values are correct).</p>
        <object id="objectContext" onload="onload_test();" data="/w3c/webperf/resources/blank_page_green.htm"></object>
        <div id="log"></div>
    </body>
</html>
