third_party/WebKit/LayoutTests/platform/win/http/tests/w3c/webperf/approved/navigation-timing/html/test_performance_attributes_exist_in_object-expected.txt

This is a testharness.js-based test.
PASS window.performance is defined 
PASS window.performance.timing is defined 
PASS window.performance.navigation is defined 
Harness: the test ran to completion.
